+++ extensions/arc/src/extension.ts
@@ -3,14 +3,14 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as arc from 'arc';
 import * as vscode from 'vscode';
-import * as loc from './localizedConstants';
 import { IconPathHelper, refreshActionId } from './constants';
+import * as loc from './localizedConstants';
+import { ConnectToControllerDialog } from './ui/dialogs/connectControllerDialog';
 import { AzureArcTreeDataProvider } from './ui/tree/azureArcTreeDataProvider';
 import { ControllerTreeNode } from './ui/tree/controllerTreeNode';
 import { TreeNode } from './ui/tree/treeNode';
-import { ConnectToControllerDialog } from './ui/dialogs/connectControllerDialog';
-import * as arc from 'arc';
 
 export async function activate(context: vscode.ExtensionContext): Promise<arc.IExtension> {
 	IconPathHelper.setExtensionContext(context);
@@ -25,7 +25,12 @@ export async function activate(context: vscode.ExtensionContext): Promise<arc.IE
 	});
 
 	vscode.commands.registerCommand('arc.connectToController', async () => {
-		await connectToController(treeDataProvider);
+		const dialog = new ConnectToControllerDialog(treeDataProvider);
+		dialog.showDialog();
+		const model = await dialog.waitForClose();
+		if (model) {
+			await treeDataProvider.addOrUpdateController(model.controllerModel, model.password);
+		}
 	});
 
 	vscode.commands.registerCommand('arc.removeController', async (controllerNode: ControllerTreeNode) => {
@@ -52,17 +57,13 @@ export async function activate(context: vscode.ExtensionContext): Promise<arc.IE
 	await checkArcDeploymentExtension();
 
 	return {
-		getAzureArcTreeDataProvider: () => treeDataProvider
-	};
-}
+		getRegisteredDataControllers: async () => {
+			return (await treeDataProvider.getChildren())
+				.filter(node => node instanceof ControllerTreeNode)
+				.map(node => (node as ControllerTreeNode).model.info);
 
-async function connectToController(treeDataProvider: AzureArcTreeDataProvider) {
-	const dialog = new ConnectToControllerDialog(treeDataProvider);
-	dialog.showDialog();
-	const model = await dialog.waitForClose();
-	if (model) {
-		await treeDataProvider.addOrUpdateController(model.controllerModel, model.password);
-	}
+		}
+	};
 }
 
 export function deactivate(): void {
@@ -71,7 +72,7 @@ export function deactivate(): void {
 async function checkArcDeploymentExtension(): Promise<void> {
 	const version = vscode.extensions.getExtension('Microsoft.arcdeployment')?.packageJSON.version;
 	if (version && version !== '0.3.2') {
-		// If we have an older verison of the deployment extension installed then uninstall it now since it's replaced
+		// If we have an older version of the deployment extension installed then uninstall it now since it's replaced
 		// by this extension. (the latest version of the Arc Deployment extension will uninstall itself so don't do
 		// anything here if that's already updated)
 		await vscode.commands.executeCommand('workbench.extensions.uninstallExtension', 'Microsoft.arcdeployment');
