+++ extensions/arc/src/models/resourceModel.ts
@@ -3,8 +3,9 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ResourceInfo } from 'arc';
 import * as vscode from 'vscode';
-import { ResourceInfo, Registration } from './controllerModel';
+import { Registration } from './controllerModel';
 
 export abstract class ResourceModel {
