+++ extensions/arc/src/ui/dashboards/controller/controllerDashboardOverviewPage.ts
@@ -3,14 +3,15 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ResourceType } from 'arc';
 import * as azdata from 'azdata';
-import * as vscode from 'vscode';
 import * as azurecore from 'azurecore';
+import * as vscode from 'vscode';
+import { getConnectionModeDisplayText, getResourceTypeIcon, parseInstanceName, resourceTypeToDisplayName } from '../../../common/utils';
+import { cssStyles, Endpoints, IconPathHelper, iconSize } from '../../../constants';
 import * as loc from '../../../localizedConstants';
-import { DashboardPage } from '../../components/dashboardPage';
-import { IconPathHelper, cssStyles, iconSize, ResourceType, Endpoints } from '../../../constants';
 import { ControllerModel } from '../../../models/controllerModel';
-import { resourceTypeToDisplayName, getResourceTypeIcon, parseInstanceName, getConnectionModeDisplayText } from '../../../common/utils';
+import { DashboardPage } from '../../components/dashboardPage';
 
 export class ControllerDashboardOverviewPage extends DashboardPage {
