+++ extensions/arc/src/ui/tree/azureArcTreeDataProvider.ts
@@ -3,12 +3,12 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ControllerInfo } from 'arc';
 import * as azdata from 'azdata';
 import * as vscode from 'vscode';
+import { ControllerModel } from '../../models/controllerModel';
 import { ControllerTreeNode } from './controllerTreeNode';
 import { TreeNode } from './treeNode';
-import { ControllerModel, ControllerInfo } from '../../models/controllerModel';
-import { IAzureArcTreeDataProvider } from 'arc';
 
 const mementoToken = 'arcControllers';
 
@@ -16,7 +16,7 @@ const mementoToken = 'arcControllers';
  * The TreeDataProvider for the Azure Arc view, which displays a list of registered
  * controllers and the resources under them.
  */
-export class AzureArcTreeDataProvider implements IAzureArcTreeDataProvider {
+export class AzureArcTreeDataProvider implements vscode.TreeDataProvider<TreeNode> {
 
 	private _credentialsProvider = azdata.credentials.getProvider('arcControllerPasswords');
 	private _onDidChangeTreeData: vscode.EventEmitter<TreeNode | undefined> = new vscode.EventEmitter<TreeNode | undefined>();
