+++ extensions/arc/src/ui/tree/postgresTreeNode.ts
@@ -3,10 +3,10 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ResourceType } from 'arc';
 import * as vscode from 'vscode';
-import { ResourceType } from '../../constants';
-import { PostgresModel } from '../../models/postgresModel';
 import { ControllerModel } from '../../models/controllerModel';
+import { PostgresModel } from '../../models/postgresModel';
 import { PostgresDashboard } from '../dashboards/postgres/postgresDashboard';
 import { ResourceTreeNode } from './resourceTreeNode';
